% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adj_mat_ineq.R
\name{adj_mat_ineq}
\alias{adj_mat_ineq}
\title{Matrix inequalities defintions
Defines inequalities from adjacency matrix input workbook}
\usage{
adj_mat_ineq(x)
}
\arguments{
\item{x}{Adjacency matrix}
}
\description{
Matrix inequalities defintions
Defines inequalities from adjacency matrix input workbook
}
