% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/centralx0.R
\name{centralx0}
\alias{centralx0}
\title{centralx0()
Solves the CENTRAL SOLUTION as the MCMC starting point in LIM::Xsample.
If central solution is not valid, defaults to LSEI}
\usage{
centralx0(full_limfile, iter = NULL, jmp = NULL, x0 = "central", ...)
}
\arguments{
\item{full_limfile}{the LIM file built from check_build()}

\item{iter}{number of iterations defined by the user.}

\item{jmp}{The jump size of the algorithm defined by the user. If NULL, jump size is internally calculated.}

\item{x0}{The starting point. Central solution.}

\item{...}{Further LIM::Xsample arguments}
}
\value{
Multiple plausible networks, starting point CENTRAL SOLUTION
}
\description{
centralx0()
Solves the CENTRAL SOLUTION as the MCMC starting point in LIM::Xsample.
If central solution is not valid, defaults to LSEI
}
