% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_build.R
\name{check_build}
\alias{check_build}
\title{multinets::check_build()
Function to check if the limfile is included or not}
\usage{
check_build(file)
}
\arguments{
\item{file}{the name of the R script containing the LIM declaration file ".R" extension}
}
\description{
multinets::check_build()
Function to check if the limfile is included or not
}
