% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defaultx0.R
\name{defaultx0}
\alias{defaultx0}
\title{defaultx0()
Solves the CENTRAL SOLUTION as the MCMC starting point in LIM::Xsample}
\usage{
defaultx0(full_limfile, iter = NULL, jmp = NULL, x0 = NULL, ...)
}
\arguments{
\item{full_limfile}{the LIM file built from check_build()}

\item{iter}{number of iterations defined by the user.}

\item{jmp}{The jump size of the algorithm defined by the user. If NULL, jump size is internally calculated.}

\item{x0}{The starting point. Defaults to LSEI.}

\item{...}{Further LIM::Xsample arguments}
}
\value{
Multiple plausible networks, starting point LSEI
}
\description{
defaultx0()
Solves the CENTRAL SOLUTION as the MCMC starting point in LIM::Xsample
}
