% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_autolimexcel.R
\name{fetch_autolimexcel}
\alias{fetch_autolimexcel}
\title{Function fetch_autolimexcel
Downloads the latest version of autoLIM-Excel from GitHub}
\usage{
fetch_autolimexcel()
}
\value{
autoLIM-Excel as .xlsx file in working directory
}
\description{
Function fetch_autolimexcel
Downloads the latest version of autoLIM-Excel from GitHub
}
