% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_def.R
\name{matrix_def}
\alias{matrix_def}
\title{matrix_def
Return matrix based on search_cols function}
\usage{
matrix_def(x, mat.type)
}
\arguments{
\item{x}{the matrix}

\item{mat.type}{the matrix type to export}
}
\description{
matrix_def
Return matrix based on search_cols function
}
