% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge.R
\name{merge}
\alias{merge}
\title{Merge
Merge lists together, name sections}
\usage{
merge(type = NULL, ...)
}
\arguments{
\item{type}{type of section}

\item{...}{other arguments}
}
\description{
Merge
Merge lists together, name sections
}
