% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_sections.R
\name{merge_sections}
\alias{merge_sections}
\title{merge_sections
Merge lists together, name sections}
\usage{
merge_sections(type = NULL, ...)
}
\arguments{
\item{type}{type of section}

\item{...}{other arguments}
}
\description{
merge_sections
Merge lists together, name sections
}
