% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meta2.R
\name{meta2}
\alias{meta2}
\title{Metadata from adjacency matrices with abbreviations
Defines metadata from adjacency matrices}
\usage{
meta2(x)
}
\arguments{
\item{x}{Adjacency matrix}
}
\description{
Metadata from adjacency matrices with abbreviations
Defines metadata from adjacency matrices
}
