% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multinets.R
\name{multinets}
\alias{multinets}
\title{MPN with multi.nets function}
\usage{
multinets(file = NULL, iter = NULL, jmp = NULL, x0 = NULL, pack = FALSE, ...)
}
\arguments{
\item{file}{the name of the R script containing the LIM declaration file ".R" extension}

\item{iter}{number of iterations defined by the user.}

\item{jmp}{The jump size of the algorithm defined by the user. If NULL, jump size is internally calculated.}

\item{x0}{The starting point. Defaults to LSEI.}

\item{pack}{Logical argument to pack the networks for analysis with package enaR}

\item{...}{Further LIM::Xsample arguments}
}
\value{
Multiple plauisble networks
}
\description{
MPN with multi.nets function
}
