% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/net_data_tidy.R
\name{net_data_tidy}
\alias{net_data_tidy}
\title{net_data_tidy
Tidies up network data input list}
\usage{
net_data_tidy(x, NLNode)
}
\arguments{
\item{x}{the sheet from network data input}

\item{NLNode}{the defined NLNodes from the main autoLIMR argument}
}
\description{
net_data_tidy
Tidies up network data input list
}
