% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_true.R
\name{pp_true}
\alias{pp_true}
\title{Primary producers True
If primary producers are included in main autoLIMR function, this function changes Q to GPP, and P to NPP}
\usage{
pp_true(x, primary_producer)
}
\arguments{
\item{x}{network input data matrix}

\item{primary_producer}{Primary producers defined in main autoLIMR function}
}
\description{
Primary producers True
If primary producers are included in main autoLIMR function, this function changes Q to GPP, and P to NPP
}
