% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepack_fun.R
\name{prepack_fun}
\alias{prepack_fun}
\title{prepack()
Function to prepack netwprks for analysis with enaR (compatibility function)}
\usage{
prepack_fun(x, limfile)
}
\arguments{
\item{x}{an object?}

\item{limfile}{the built limfile}
}
\description{
prepack()
Function to prepack netwprks for analysis with enaR (compatibility function)
}
