% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pvar_wo_ex.R
\name{pvar_wo_ex}
\alias{pvar_wo_ex}
\title{Production variable definition without exports}
\usage{
pvar_wo_ex(x, respiration, NLNode)
}
\arguments{
\item{x}{matrices defined from Network Data Input workbook}

\item{respiration}{If respiration = TRUE in main autoLIMR argument}

\item{NLNode}{the defined NLNodes from the main autoLIMR argument}
}
\description{
Production variable definition without exports
}
