% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qvar.R
\name{qvar}
\alias{qvar}
\title{Consumption (Q) Variable definition
Living nodes only, depends on Inputs}
\usage{
qvar(x)
}
\arguments{
\item{x}{network input data matrix}
}
\description{
Consumption (Q) Variable definition
Living nodes only, depends on Inputs
}
