% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_cols.R
\name{search_cols}
\alias{search_cols}
\title{search_cols
search and return columns that match type (as character vector) "Inputs", "Exports", "Assimilation Efficieny"}
\usage{
search_cols(x, col.match)
}
\arguments{
\item{x}{the matrix}

\item{col.match}{the node type to match}
}
\description{
search_cols
search and return columns that match type (as character vector) "Inputs", "Exports", "Assimilation Efficieny"
}
