% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uvar.R
\name{uvar}
\alias{uvar}
\title{Unused material/energy variable definition}
\usage{
uvar(x)
}
\arguments{
\item{x}{matrices defined from Network Data Input workbook}
}
\description{
Unused material/energy variable definition
}
